// frontend/src/api/generated/model/mapPointResponse.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Krisefikser API
 * API for the Krisefikser application
 * OpenAPI spec version: 1.0
 */
import type { MapPointTypeResponse } from './mapPointTypeResponse'

export interface MapPointResponse {
  id: number
  latitude: number
  longitude: number
  type: MapPointTypeResponse
}